backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
pillow==11.0.0
psycopg2-binary==2.9.10
httpx==0.27.0
python-multipart==0.0.19
